+++ src/main/java/com/fiap/lanchonete/domain/pojo/CreateAlimentoDto.java
@@ -3,7 +3,7 @@ package com.fiap.lanchonete.domain.pojo;
 import jakarta.validation.constraints.NotEmpty;
 
 public record CreateAlimentoDto(
-    @NotEmpty Integer codigoTipoAlimento,
+    @NotEmpty Short codigoTipoAlimento,
     @NotEmpty String nomeAlimento,
     @NotEmpty String nomeFunciAlter) {
 }
