+++ src/main/java/com/fiap/lanchonete/domain/pojo/DeleteAlimentoDto.java
@@ -1,4 +1,4 @@
 package com.fiap.lanchonete.domain.pojo;
 
-public record DeleteAlimentoDto(Integer codigoAlimento, Integer codigoTipoAlimento) {
+public record DeleteAlimentoDto(Short codigoAlimento, Short codigoTipoAlimento) {
 }
