+++ src/main/java/com/fiap/lanchonete/domain/pojo/EditAlimentoDto.java
@@ -1,8 +1,8 @@
 package com.fiap.lanchonete.domain.pojo;
 
 public record EditAlimentoDto(
-    Integer codigoTipoAlimento,
-    Integer codigoAlimento,
+    Short codigoTipoAlimento,
+    Short codigoAlimento,
     String nomeAlimento,
     String nomeFunciAlter) {
 }
